python/requirements.txt
boto3==1.12.35
psycopg2-binary==2.8.4
aiopg==1.0.0
